requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.1
Flask-Caching==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
numpy==1.24.2
pythondialog==3.4.0
requests==2.22.0
sqlparse==0.4.4
Werkzeug==3.0.1
pandas==2.0.3
tk
google-cloud-storage
